# internal/user/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Transaction API.",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1/api",
    "paths": {
        "/user/register": {
            "post": {
                "description": "register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register register",
                "parameters": [
                    {
                        "description": "register screen",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_khuong02_backend_internal_user_payload.Register"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_khuong02_backend_internal_user_dtos.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_khuong02_backend_pkg_helper.ResponseErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_khuong02_backend_pkg_helper.ResponseErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_khuong02_backend_internal_user_dtos.AuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "github_com_khuong02_backend_internal_user_payload.Register": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_khuong02_backend_pkg_helper.ResponseErr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {}
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {}
            }
        }
    },
    "securityDefinitions": {
        "AuthToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}